move download method to httputils
